3. inner class to upper class
